Implement new commands add/del-builders and add/del-workers
These take a json file argument and start/stop builders/testworkers
as requested
Adds a "oneshot" builder and tester parameter - when encountered,
such a thread would only be kept alive until there are actual
waiting items in the respective queue.
This is good for overflow threads that are only spun up when there's
a big buildup of items in the queue to help chew through them
faster and then exit once the backlog was completed.